DIY buzzer code and minor edits to save space
shaving serial messages

remember, no virtual methods

array of settings for input sensors

hard coded controller name

intermittent buzzer

osPID_serial.ino edits

buzzer pin constants

serialSpeedTable/100

turn off function inlining

squeezed osPID_Menu::drawNotificationCursor

+++ osPID_Firmware/ospDecimalValue.h
@@ -263,7 +263,7 @@ template<int lDEC, int rDEC> inline bool operator >= (const ospDecimalValue<lDEC
 
 template<int D> inline ospDecimalValue<D> makeDecimal(double val)
 {
-  return (ospDecimalValue<D>) {(int)(val * ospPow10<D>::value + 0.5)}; // rounded
+  return (ospDecimalValue<D>) {(int)(val * (double)ospPow10<D>::value + 0.5)}; // rounded
 }
 
 #endif
